Move var to separate file
